# public/docs/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "Taklimakan cryptodata",
    "version": "v0.0.1"
  },
  "paths": {
...
    "/api/v2/coin/history": {
      "get": {
        "tags": [
          "v2"
        ],
        "summary": "Coin history ",
        "description": "Coin history",
        "operationId": "getChartsData",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "What currency symbol",
            "required": true,
            "type": "string"
          },
          {
            "name": "resolution",
            "in": "query",
            "description": "What currency resolution",
            "required": true,
            "type": "string"
          },
          {
            "name": "from",
            "in": "query",
            "description": "What currency from date",
            "required": true,
            "type": "string"
          },
          {
            "name": "to",
            "in": "query",
            "description": "What currency to date",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "": {
                  "type": "object",
                  "properties": {
                    "c": {
                      "type": "array"
                    },
                    "h": {
                      "type": "array"
                    },
                    "l": {
                      "type": "array"
                    },
                    "o": {
                      "type": "array"
                    },
                    "t": {
                      "type": "array"
                    },
                    "v": {
                      "type": "array"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request"
            }
          }
        }
      }
    },
    "/api/v2/coin/symbols": {
      "get": {
        "tags": [
          "v2"
        ],
        "summary": "Coin symbols ",
        "description": "Coin symbols",
        "operationId": "getChartsData",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "What currency symbol",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "": {
                  "type": "object",
                  "properties": {
                    "description": {
                      "type": "string"
                    },
                    "exchange-listed": {
                      "type": "string"
                    },
                    "exchange-traded": {
                      "type": "string"
                    },
                    "has_intraday": {
                      "type": "boolean"
                    },
                    "has_no_volume": {
                      "type": "boolean"
                    },
                    "minmovement": {
                      "type": "number"
                    },
                    "minmovement2": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "pointvalue": {
                      "type": "number"
                    },
                    "pricescale": {
                      "type": "number"
                    },
                    "session": {
                      "type": "string"
                    },
                    "ticker": {
                      "type": "string"
                    },
                    "timezone": {
                      "type": "string"
                    },
                    "has_daily": {
                      "type": "boolean"
                    },
                    "has_weekly_and_monthly": {
                      "type": "boolean"
                    },
                    "type": {
                      "type": "string"
                    },
                    "supported_resolutions": {
                      "type": "array"
                    },
                    "intraday_multipliers": {
                      "type": "array"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          }
        }
      }
    },
...
    "/api/coin/favorite": {
      "get": {
        "tags": [
          "coin"
        ],
        "summary": "Get favorite data ",
        "description": "Data favorite",
        "operationId": "getChartsData",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "Coin symbol",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "pagination": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "number"
                    },
                    "perPage": {
                      "type": "number"
                    },
                    "skip": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "coins": {
                      "type": "string"
                    }
                  }
                },
                "data": {
                  "type": "array",
                  "properties": {
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request"
            }
          }
        }
      }
    },
    "/api/coin/top_market_pairs": {
      "get": {
        "tags": [
          "coin"
        ],
        "summary": "Get favorite data ",
        "description": "Data favorite",
        "operationId": "getChartsData",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "Coin symbol",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "pagination": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "number"
                    },
                    "perPage": {
                      "type": "number"
                    },
                    "skip": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "symbol": {
                      "type": "string"
                    },
                    "cryptocurrency_id": {
                      "type": "number"
                    }
                  }
                },
                "data": {
                  "type": "array",
                  "properties": {
                    "rank": {
                      "type": "number"
                    },
                    "logo": {
                      "type": "string"
                    },
                    "exchange_name": {
                      "type": "string"
                    },
                    "pair_name": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "volume_24": {
                      "type": "number"
                    },
                    "percent_volume_24": {
                      "type": "number"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request"
            }
          }
        }
      }
    },
...
    "/api/coin/currency": {
      "get": {
        "tags": [
          "coin"
        ],
        "summary": "Get currency data ",
        "description": "Data currency",
        "operationId": "getChartsData",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "What currency symbol",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "query",
            "description": "What crypto id",
            "required": false,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "required": [
                "name"
              ],
              "properties": {
                "status": {
                  "type": "object",
                  "properties": {
                    "error_message": {
                      "type": "string"
                    },
                    "error_code": {
                      "type": "integer"
                    }
                  }
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "cryptocurrency_id": {
                      "type": "number"
                    },
                    "name": {
                      "type": "string"
                    },
                    "symbol": {
                      "type": "string"
                    },
                    "circulating_supply": {
                      "type": "number"
                    },
                    "max_supply": {
                      "type": "number"
                    },
                    "num_market_pairs": {
                      "type": "number"
                    },
                    "total_supply": {
                      "type": "number"
                    },
                    "logo": {
                      "type": "string"
                    },
                    "percent_changes_weekly": {
                      "type": "number"
                    },
                    "percent_changes_monthly": {
                      "type": "number"
                    },
                    "percent_changes_three_month": {
                      "type": "number"
                    },
                    "description": {
                      "type": "string"
                    },
                    "url": {
                      "type": "object",
                      "website": {
                        "type": "array"
                      },
                      "twitter": {
                        "type": "array"
                      },
                      "reddit": {
                        "type": "array"
                      },
                      "message_board": {
                        "type": "array"
                      },
                      "announcement": {
                        "type": "array"
                      },
                      "chat": {
                        "type": "array"
                      },
                      "explorer": {
                        "type": "array"
                      },
                      "source_code": {
                        "type": "array"
                      }
                    },
                    "circulating_supply_percent": {
                      "type": "number"
                    },
                    "market_cap": {
                      "type": "number"
                    },
                    "usd_price": {
                      "type": "number"
                    },
                    "volume_24": {
                      "type": "number"
                    },
                    "percent_change_price": {
                      "type": "number"
                    },
                    "btc_volume_24": {
                      "type": "number"
                    },
                    "btc_market_cap": {
                      "type": "number"
                    },
                    "btc_percent_change": {
                      "type": "number"
                    },
                    "rank": {
                      "type": "number"
                    }
                  }
                },
                "filters": {
                  "type": "object",
                  "properties": {
                    "cryptocurrency_id": {
                      "type": "number"
                    },
                    "symbol": {
                      "type": "string"
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid request"
            }
          }
        }
      }
    },
...
      },
    "consumes": [
      "application/json"
    ]
  }